Update installation instructions.
The current version of Middleman copies over all folders, including the .git
folder when running the "init" command. If this command is run on an existing
git repo, the .git folder will be overwritten. The current Middleman master
version fixes this, so as a workaround until it is released, have users
download the tarball instead.